requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-chroma
chromadb
streamlit
pandas
dask[dataframe]
ollama
sentence-transformers
watchdog
rank_bm25
prometheus_client
langchain-huggingface
langchain-core
langchain-text-splitters
pymupdf
openpyxl
python-docx
